Refactor long screens into reusable widgets
